src/bin/tectonic: split main compilation implementation to a separate module
